seperate lock class into own file
